[refactor] Set most helper types to be nested, unify jumps style for cycles in parsing algorithms.

+++ src/Lexing/Tokens/UnitToken.cs
@@ -3,11 +3,11 @@ namespace NMLServer.Lexing.Tokens;
 internal sealed class UnitToken : Token
 {
     private UnitType _unit;
-    public byte length { get; }
+    public ushort length { get; }
 
     public UnitToken(int start, UnitType value, int length) : base(start)
     {
         _unit = value;
-        this.length = (byte)length;
+        this.length = (ushort)length;
     }
 }
